backend/pyproject.toml
[tool.poetry]
name = "colorization-backend"
version = "0.1.0"
description = "Backend API for image colorization"
authors = ["Your Team <team@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pillow = "^10.1.0"
python-multipart = "^0.0.6"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
aiofiles = "^23.2.1"
loguru = "^0.7.2"
torch = "^2.1.0"
torchvision = "^0.16.0"
fastai = "1.0.61"
fastprogress = "^1.0.3"
matplotlib = "^3.8.0"
numpy = "^1.26.0"
scikit-image = "^0.22.0"
deoldify = "^0.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"